Move a method used everywhere out of `intrinsic.rs`
